requirements.txt
boto3==1.19.3
botocore==1.22.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
oauthlib==3.1.1
py-mini-racer==0.6.0
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.2
python-dotenv==0.19.0
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
